pyproject.toml
[tool.poetry]
name = "shelfsense-ai"
version = "1.0.0"
description = "Vision-Language Powered Personal Library Intelligence"
authors = ["ShelfSense AI Team"]
readme = "README.md"
license = "MIT"
packages = [{include = "shelfsense"}]

[tool.poetry.dependencies]
python = "^3.10"

# Core ML/DL
torch = "^2.1.0"
torchvision = "^0.16.0"
ultralytics = "^8.0.200"  # YOLOv8
transformers = "^4.35.0"
sentence-transformers = "^2.2.2"

# Computer Vision
opencv-python = "^4.8.1"
Pillow = "^10.1.0"
albumentations = "^1.3.1"

# OCR
easyocr = "^1.7.1"
pytesseract = "^0.3.10"

# Vector Store
faiss-cpu = "^1.7.4"
# faiss-gpu = "^1.7.4"  # Uncomment for GPU support

# Web Framework
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
python-multipart = "^0.0.6"

# Database
sqlalchemy = "^2.0.23"
asyncpg = "^0.29.0"
alembic = "^1.12.1"
psycopg2-binary = "^2.9.9"

# Data Validation
pydantic = "^2.5.2"
pydantic-settings = "^2.1.0"

# HTTP Client
httpx = "^0.25.2"
aiohttp = "^3.9.1"

# LLM Integration
openai = "^1.3.0"
anthropic = "^0.7.0"

# Text Processing
rank-bm25 = "^0.2.2"
nltk = "^3.8.1"
Unidecode = "^1.3.7"

# Utilities
numpy = "^1.26.2"
pandas = "^2.1.3"
tqdm = "^4.66.1"
python-dotenv = "^1.0.0"
loguru = "^0.7.2"
tenacity = "^8.2.3"

# Caching
cachetools = "^5.3.2"
redis = "^5.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.11.0"
ruff = "^0.1.6"
mypy = "^1.7.0"
pre-commit = "^3.6.0"
jupyter = "^1.0.0"
ipywidgets = "^8.1.1"
matplotlib = "^3.8.2"
seaborn = "^0.13.0"

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
